requirements.txt
# requirements.txt
streamlit
numpy
opencv-python-headless
scikit-learn
Pillow
rembg[cpu]